fill memory with constant byte

+++ 0x07-pointers_arrays_strings/0-memset.c
@@ -13,7 +13,7 @@
 
 char *_memset(char *s, char b, unsigned int n)
 {
-	memset(s, (int)b, n);
+	memset(s, b, n);
 
 	return (s);
 }
